:sparkles: feat : study와 book & lecture 매핑
resolves : #72

+++ ZZANDI/src/main/java/com/ll/zzandi/repository/LectureRepository.java
@@ -5,6 +5,6 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface LectureRepository extends JpaRepository<Lecture, Integer> {
+public interface LectureRepository extends JpaRepository<Lecture, Long> {
     Optional<Lecture> findById(Long id);
 }
